+++ Civi/Api4/CampagnodonCivirulesLog.php
@@ -1,0 +1,13 @@
+<?php
+namespace Civi\Api4;
+
+/**
+ * CampagnodonCivirulesLog entity.
+ *
+ * Provided by the Campagnodon extension.
+ *
+ * @package Civi\Api4
+ */
+class CampagnodonCivirulesLog extends Generic\DAOEntity {
+
+}